Update AnkiPoetry.py
Updated to display the stanza break in the cards after the last line of a stanza. The <br> is appended to the last line of stanzas; an alternative version of this prepends the <br> to the first line of stanzas. I think I prefer this way, but jury is out.